Split CLI Commands into Package
That one module was getting really crowded.
